OCR_LLM_App/requirements.txt
streamlit
numpy
pillow
easyocr
requests
tensorflow